Changed API to exchangerate.host

+++ CurrencyConverterCore/Services/IExchangeRateService.cs
@@ -5,6 +5,6 @@ namespace CurrencyConverterCore.Services
     public interface IExchangeRateService
     {
         Task<List<Currency>> GetCurrenciesNames();
-        Task<double> GetCurrencyRates(double amount, string from, string to);
+        Task<double> GetCurrencyRates(decimal amount, string from, string to);
     }
 }
